src/wizard_ops/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.9.1",
    "uvicorn==0.34.0",
    "fastapi==0.115.6",
    "google-cloud-storage>=3.8.0",
]


[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]
torchvision = [{ index = "pytorch-cpu" }]
torchmetrics = [{ index = "pytorch-cpu" }]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv]
override-dependencies = ["torch==2.10.0+cpu", "torchvision==0.25.0+cpu"]
